Meeting room made larger
Table made bigger, more chairs placed and a few extra cups on the table for the cosmetic view. Of course the meeting room itself was also made bigger so that a Jitsi meeting can be attended from a larger area. The markings on the floor were also adjusted to this new size.